requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
requests==2.31.0
psutil==5.9.6
python-json-logger==2.0.7
elasticsearch==8.11.0
alembic==1.12.1
SQLAlchemy==2.0.23
